requirements.txt
python-dotenv==0.20.0
requests==2.28.1
coverage==6.4.2
flask-restx==0.5.1
flask-mongoengine==1.0.0